@@ -29,7 +29,7 @@
     "doz": "dozjs/doz#v5"
   },
   "devDependencies": {
-    "webpack": "4.41.4",
-    "webpack-cli": "3.3.10"
+    "webpack": "5.76.0",
+    "webpack-cli": "5.0.1"
   }
 }
