@@ -23,13 +23,13 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@babel/core": "7.12.3",
-    "@babel/plugin-transform-react-jsx": "7.14.9",
-    "@babel/preset-env": "7.12.1",
-    "babel-loader": "8.1.0",
-    "terser-webpack-plugin": "1.3.0",
-    "webpack": "5.51.1",
-    "webpack-cli": "4.8.0"
+    "@babel/core": "7.21.0",
+    "@babel/plugin-transform-react-jsx": "7.21.0",
+    "@babel/preset-env": "7.20.2",
+    "babel-loader": "9.1.2",
+    "terser-webpack-plugin": "5.3.7",
+    "webpack": "5.76.0",
+    "webpack-cli": "5.0.1"
   },
   "dependencies": {
     "fre": "2.5.5"
