@@ -23,16 +23,14 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@webpack-cli/init": "0.1.3",
     "babel-plugin-syntax-dynamic-import": "6.18.0",
-    "terser-webpack-plugin": "1.2.3",
-    "uglifyjs-webpack-plugin": "2.1.2",
-    "webpack": "4.29.6",
-    "webpack-cli": "3.2.3"
+    "terser-webpack-plugin": "5.3.7",
+    "webpack": "5.76.0",
+    "webpack-cli": "5.0.1"
   },
   "dependencies": {
-    "@hullo/browser": "0.8.2",
-    "@hullo/core": "0.8.3",
-    "@hullo/dom": "0.8.3"
+    "@hullo/browser": "0.9.0",
+    "@hullo/core": "0.9.0",
+    "@hullo/dom": "0.9.0"
   }
 }
