@@ -27,15 +27,15 @@
   },
   "homepage": "https://github.com/krausest/js-framework-benchmark#readme",
   "dependencies": {
-    "lit": "^2.1.1"
+    "lit": "2.6.1"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@rollup/plugin-node-resolve": "11.2.1",
     "rollup": "2.45.2",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
+    "rollup-plugin-minify-html-literals": "1.2.6",
     "rollup-plugin-terser": "7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4"
+    "rollup-plugin-typescript2": "0.34.1",
+    "tslib": "2.5.0",
+    "typescript": "4.9.5"
   }
 }
