@@ -9,16 +9,16 @@
   "author": "Alexey Oganezov <alexey@oganezov.work>",
   "license": "MIT",
   "devDependencies": {
-    "ts-loader": "9.1.2",
-    "typescript": "4.2.4",
-    "webpack": "5.37.0",
-    "webpack-cli": "4.7.0"
+    "ts-loader": "9.4.2",
+    "typescript": "4.9.5",
+    "webpack": "5.76.0",
+    "webpack-cli": "5.0.1"
   },
   "js-framework-benchmark": {
     "frameworkVersionFromPackage": "oldskull",
     "frameworkHomeURL": "https://github.com/alexeyoganezov/oldskull"
   },
   "dependencies": {
-    "oldskull": "0.1.2"
+    "oldskull": "2.0.0"
   }
 }
