@@ -7,8 +7,8 @@
     "frameworkHomeURL": "https://ractive.js.org/"
   },
   "scripts": {
-    "build-dev": "webpack -w -d",
-    "build-prod": "webpack -p"
+    "build-dev": "webpack --watch",
+    "build-prod": "webpack --mode production"
   },
   "keywords": [
     "ractive"
@@ -21,12 +21,12 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "babel-loader": "8.0.6",
+    "babel-loader": "9.1.2",
     "ractive-bin-loader": "^1",
-    "webpack": "4.34.0",
-    "webpack-cli": "3.3.4",
-    "@babel/core": "7.4.5",
-    "@babel/preset-env": "7.4.5"
+    "webpack": "5.76.1",
+    "webpack-cli": "5.0.1",
+    "@babel/core": "7.21.0",
+    "@babel/preset-env": "7.20.2"
   },
   "dependencies": {
     "ractive": "1.3.6"
