@@ -23,14 +23,14 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@babel/core": "7.1.2",
-    "@babel/preset-env": "7.1.0",
-    "@babel/preset-react": "7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.1.0",
-    "babel-loader": "8.0.4",
-    "terser-webpack-plugin": "1.1.0",
-    "webpack": "4.22.0",
-    "webpack-cli": "3.1.2"
+    "@babel/core": "7.21.0",
+    "@babel/preset-env": "7.20.2",
+    "@babel/preset-react": "7.18.6",
+    "@babel/plugin-proposal-class-properties": "7.18.6",
+    "babel-loader": "9.1.2",
+    "terser-webpack-plugin": "5.3.7",
+    "webpack": "5.76.1",
+    "webpack-cli": "5.0.1"
   },
   "dependencies": {
     "rax": "0.6.7"
