@@ -1,31 +1,29 @@
 'use strict';
 var path = require('path')
-const MinifyPlugin = require("babel-minify-webpack-plugin");
 var webpack = require('webpack')
+const TerserPlugin = require('terser-webpack-plugin');
 
-var cache = {};
-var loaders = [
-    {
-        test: /\.jsx$/,
-        loader: 'babel-loader'
-    },
-    {
-        test: /\.es6\.js$/,
-        loader: 'babel-loader'
-    },
-    {
-        test: /\.css$/,
-        loader: 'style-loader!css-loader'
-    }
-];
 var extensions = [
     '.js', '.jsx', '.es6.js', '.msx'
 ];
 
 module.exports = [{
-    cache: cache,
+    cache: true,
     module: {
-        rules: loaders
+        rules: [
+            {
+                test: /\.jsx$/,
+                use: {
+                    loader: 'babel-loader'
+                }
+            },
+            {
+                test: /\.es6\.js$/,
+                use: {
+                    loader: 'babel-loader'
+                }
+            }
+        ]
     },
     entry: {
         main: './src/Main.jsx',
@@ -51,7 +49,12 @@ module.exports = [{
             'process.env': {
                 'NODE_ENV': JSON.stringify('production')
             }
-        }),
-        new MinifyPlugin()
-    ]
+        })
+    ],
+	optimization: {
+        minimize: true,
+		minimizer: [
+			new TerserPlugin({}),
+		]
+	},    
 }];
